Moved details page content to Settings tab Added Repositories tab
