+++ hub/src/components/index.ts
@@ -28,6 +28,8 @@ export type { AutosuggestProps } from '@cloudscape-design/components/autosuggest
 export { default as Grid } from '@cloudscape-design/components/grid';
 export { default as HelpPanel } from '@cloudscape-design/components/help-panel';
 export type { HelpPanelProps } from '@cloudscape-design/components/help-panel';
+export { default as Tabs } from '@cloudscape-design/components/tabs';
+export type { TabsProps } from '@cloudscape-design/components/tabs';
 
 // custom components
 export { NavigateLink } from './NavigateLink';
