+++ hub/src/pages/Project/Details/Repositories/index.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import { useParams } from 'react-router-dom';
+
+import { useBreadcrumbs } from 'hooks';
+import { ROUTES } from 'routes';
+
+export const ProjectRepositories: React.FC = () => {
+    const { t } = useTranslation();
+    const params = useParams();
+    const paramProjectName = params.name ?? '';
+
+    useBreadcrumbs([
+        {
+            text: t('navigation.projects'),
+            href: ROUTES.PROJECT.LIST,
+        },
+        {
+            text: paramProjectName,
+            href: ROUTES.PROJECT.DETAILS.REPOSITORIES.FORMAT(paramProjectName),
+        },
+        {
+            text: t('projects.repositories'),
+            href: ROUTES.PROJECT.DETAILS.SETTINGS.FORMAT(paramProjectName),
+        },
+    ]);
+
+    return <div></div>;
+};
